backend/requirements.txt
Flask
Flask-Cors
spotipy
python-dotenv
gunicorn
